// F-Learning/Frontend/F-LearningFE/src/app/features/manage-course/interface/manage-course.interface.ts
export interface Course {
    id?: number; // Optional as it might be generated by the server
    title: string;
    description: string;
    price: number;
    thumbnailUrl: string;
    status?: string; // Optional status
    categoryId: number;
    categoryName?: string; // Optional as it might be returned by the server
    lessonCount?: number; // Optional lesson count
  }